Rename commonDefs.hpp to commonDefs.h
- This commit also moves the FIM db queries to the fimDB.hpp (this avoid
  including c++ code in c.
- Also changes the handlers to a shared pointer.
