--- src/syscheckd/db/include/commonDefs.hpp
@@ -1,215 +1,0 @@
-/*
- * @file db_statement.hpp
- * @brief Definition of FIM database statements.
- * @date 2021-09-06
- *
- * @copyright Copyright (C) 2015-2021 Wazuh, Inc.
- */
-
-#ifndef DB_COMMONDEFS_HPP
-#define DB_COMMONDEFS_HPP
-// #include "logging_helper.h"
-// typedef void((*log_callback_t)(const syscollector_log_level_t level, const char* log));
-typedef void((*fim_sync_callback_t)(const void* buffer));
-
-constexpr auto DATABASE_TEMP {"queue/fim/db/fim_dbsync.db"};
-
-constexpr auto CREATE_FILE_DB_STATEMENT
-{
-    R"(CREATE TABLE IF NOT EXISTS file_entry (
-    path TEXT NOT NULL,
-    mode INTEGER,
-    last_event INTEGER,
-    scanned INTEGER,
-    options INTEGER,
-    checksum TEXT NOT NULL,
-    dev INTEGER,
-    inode INTEGER,
-    size INTEGER,
-    perm TEXT,
-    attributes TEXT,
-    uid INTEGER,
-    gid INTEGER,
-    user_name TEXT,
-    group_name TEXT,
-    hash_md5 TEXT,
-    hash_sha1 TEXT,
-    hash_sha256 TEXT,
-    mtime INTEGER,
-    PRIMARY KEY(path));)"
-};
-
-constexpr auto CREATE_REGISTRY_KEY_DB_STATEMENT
-{
-    R"(CREATE TABLE IF NOT EXISTS registry_key (
-    path TEXT NOT NULL,
-    perm TEXT,
-    uid INTEGER,
-    gid INTEGER,
-    user_name TEXT,
-    group_name TEXT,
-    mtime INTEGER,
-    arch TEXT CHECK (arch IN ('[x32]', '[x64]')),
-    scanned INTEGER,
-    last_event INTEGER,
-    checksum TEXT NOT NULL,
-    item_id TEXT,
-    PRIMARY KEY (arch, path));)"
-};
-static const std::vector<std::string> REGISTRY_KEY_ITEM_ID_FIELDS{"arch", "path"};
-
-constexpr auto CREATE_REGISTRY_VALUE_DB_STATEMENT
-{
-    R"(CREATE TABLE IF NOT EXISTS registry_data (
-    key_id INTEGER,
-    name TEXT,
-    type INTEGER,
-    size INTEGER,
-    hash_md5 TEXT,
-    hash_sha1 TEXT,
-    hash_sha256 TEXT,
-    scanned INTEGER,
-    last_event INTEGER,
-    checksum TEXT NOT NULL,
-    item_id TEXT,
-    PRIMARY KEY(key_id, name)
-    FOREIGN KEY (key_id) REFERENCES registry_key(item_id));)"
-};
-static const std::vector<std::string> REGISTRY_VALUE_ITEM_ID_FIELDS{"key_id", "name"};
-
-constexpr auto FIM_FILE_SYNC_CONFIG_STATEMENT
-{
-    R"(
-    {
-        "decoder_type":"JSON_RANGE",
-        "table":"file_entry",
-        "component":"fim_file_sync",
-        "index":"path",
-        "checksum_field":"checksum",
-        "no_data_query_json": {
-                "row_filter":"WHERE path BETWEEN '?' and '?' ORDER BY path",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "count_range_query_json": {
-                "row_filter":"WHERE path BETWEEN '?' and '?' ORDER BY path",
-                "count_field_name":"count",
-                "column_list":["count(*) AS count "],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "row_data_query_json": {
-                "row_filter":"WHERE path ='?'",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "range_checksum_query_json": {
-                "row_filter":"WHERE path BETWEEN '?' and '?' ORDER BY path",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        }
-    }
-    )"
-};
-
-constexpr auto FIM_REGISTRY_SYNC_CONFIG_STATEMENT
-{
-    R"(
-    {
-        "decoder_type":"JSON_RANGE",
-        "table":"registry_key",
-        "component":"fim_registry_sync",
-        "index":"item_id",
-        "checksum_field":"checksum",
-        "no_data_query_json": {
-                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "count_range_query_json": {
-                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
-                "count_field_name":"count",
-                "column_list":["count(*) AS count "],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "row_data_query_json": {
-                "row_filter":"WHERE item_id ='?'",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "range_checksum_query_json": {
-                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        }
-    }
-    )"
-};
-
-constexpr auto FIM_VALUE_SYNC_CONFIG_STATEMENT
-{
-    R"(
-    {
-        "decoder_type":"JSON_RANGE",
-        "table":"registry_data",
-        "component":"fim_value_sync",
-        "index":"item_id",
-        "checksum_field":"checksum",
-        "no_data_query_json": {
-                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "count_range_query_json": {
-                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
-                "count_field_name":"count",
-                "column_list":["count(*) AS count "],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "row_data_query_json": {
-                "row_filter":"WHERE item_id ='?'",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        },
-        "range_checksum_query_json": {
-                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
-                "column_list":["*"],
-                "distinct_opt":false,
-                "order_by_opt":""
-        }
-    }
-    )"
-};
-
-/* Statement related to files items. Defines everything necessary to perform the synchronization loop */
-constexpr auto FIM_FILE_START_CONFIG_STATEMENT
-{
-    R"({"table":"file_entry"})"
-    //TO DO
-};
-
-/* Statement related to registries items. Defines everything necessary to perform the synchronization loop */
-constexpr auto FIM_REGISTRY_START_CONFIG_STATEMENT
-{
-    R"({"table":"registry_key"})"
-    //TO DO
-};
-
-/* Statement related to values items. Defines everything necessary to perform the synchronization loop */
-constexpr auto FIM_VALUE_START_CONFIG_STATEMENT
-{
-    R"({"table":"registry_data"})"
-    //TO DO
-};
-
-#endif // DB_STATEMENT_HPP
