+++ src/syscheckd/db/include/fimDB.hpp
@@ -13,7 +13,7 @@
 #define _FIMDB_HPP
 #include "dbsync.hpp"
 #include "rsync.hpp"
-#include "commonDefs.hpp"
+#include "commonDefs.h"
 #include <condition_variable>
 #include <mutex>
 
@@ -24,15 +24,204 @@ extern "C"
 }
 #endif
 
-typedef void((*send_data_callback_t)(const char* log, const char* tag));
-typedef void((*logging_callback_t)(modules_log_level_t level, const char* tag));
-
-enum class dbQueryResult
+constexpr auto CREATE_FILE_DB_STATEMENT
 {
-    SUCCESS,
-    MAX_ROWS_ERROR,
-    DBSYNC_ERROR
+    R"(CREATE TABLE IF NOT EXISTS file_entry (
+    path TEXT NOT NULL,
+    mode INTEGER,
+    last_event INTEGER,
+    scanned INTEGER,
+    options INTEGER,
+    checksum TEXT NOT NULL,
+    dev INTEGER,
+    inode INTEGER,
+    size INTEGER,
+    perm TEXT,
+    attributes TEXT,
+    uid INTEGER,
+    gid INTEGER,
+    user_name TEXT,
+    group_name TEXT,
+    hash_md5 TEXT,
+    hash_sha1 TEXT,
+    hash_sha256 TEXT,
+    mtime INTEGER,
+    PRIMARY KEY(path));)"
 };
+
+constexpr auto CREATE_REGISTRY_KEY_DB_STATEMENT
+{
+    R"(CREATE TABLE IF NOT EXISTS registry_key (
+    path TEXT NOT NULL,
+    perm TEXT,
+    uid INTEGER,
+    gid INTEGER,
+    user_name TEXT,
+    group_name TEXT,
+    mtime INTEGER,
+    arch TEXT CHECK (arch IN ('[x32]', '[x64]')),
+    scanned INTEGER,
+    last_event INTEGER,
+    checksum TEXT NOT NULL,
+    item_id TEXT,
+    PRIMARY KEY (arch, path));)"
+};
+static const std::vector<std::string> REGISTRY_KEY_ITEM_ID_FIELDS{"arch", "path"};
+
+constexpr auto CREATE_REGISTRY_VALUE_DB_STATEMENT
+{
+    R"(CREATE TABLE IF NOT EXISTS registry_data (
+    key_id INTEGER,
+    name TEXT,
+    type INTEGER,
+    size INTEGER,
+    hash_md5 TEXT,
+    hash_sha1 TEXT,
+    hash_sha256 TEXT,
+    scanned INTEGER,
+    last_event INTEGER,
+    checksum TEXT NOT NULL,
+    item_id TEXT,
+    PRIMARY KEY(key_id, name)
+    FOREIGN KEY (key_id) REFERENCES registry_key(item_id));)"
+};
+static const std::vector<std::string> REGISTRY_VALUE_ITEM_ID_FIELDS{"key_id", "name"};
+
+constexpr auto FIM_FILE_SYNC_CONFIG_STATEMENT
+{
+    R"(
+    {
+        "decoder_type":"JSON_RANGE",
+        "table":"file_entry",
+        "component":"fim_file_sync",
+        "index":"path",
+        "checksum_field":"checksum",
+        "no_data_query_json": {
+                "row_filter":"WHERE path BETWEEN '?' and '?' ORDER BY path",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "count_range_query_json": {
+                "row_filter":"WHERE path BETWEEN '?' and '?' ORDER BY path",
+                "count_field_name":"count",
+                "column_list":["count(*) AS count "],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "row_data_query_json": {
+                "row_filter":"WHERE path ='?'",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "range_checksum_query_json": {
+                "row_filter":"WHERE path BETWEEN '?' and '?' ORDER BY path",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        }
+    }
+    )"
+};
+
+constexpr auto FIM_REGISTRY_SYNC_CONFIG_STATEMENT
+{
+    R"(
+    {
+        "decoder_type":"JSON_RANGE",
+        "table":"registry_key",
+        "component":"fim_registry_sync",
+        "index":"item_id",
+        "checksum_field":"checksum",
+        "no_data_query_json": {
+                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "count_range_query_json": {
+                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
+                "count_field_name":"count",
+                "column_list":["count(*) AS count "],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "row_data_query_json": {
+                "row_filter":"WHERE item_id ='?'",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "range_checksum_query_json": {
+                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        }
+    }
+    )"
+};
+
+constexpr auto FIM_VALUE_SYNC_CONFIG_STATEMENT
+{
+    R"(
+    {
+        "decoder_type":"JSON_RANGE",
+        "table":"registry_data",
+        "component":"fim_value_sync",
+        "index":"item_id",
+        "checksum_field":"checksum",
+        "no_data_query_json": {
+                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "count_range_query_json": {
+                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
+                "count_field_name":"count",
+                "column_list":["count(*) AS count "],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "row_data_query_json": {
+                "row_filter":"WHERE item_id ='?'",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        },
+        "range_checksum_query_json": {
+                "row_filter":"WHERE item_id BETWEEN '?' and '?' ORDER BY item_id",
+                "column_list":["*"],
+                "distinct_opt":false,
+                "order_by_opt":""
+        }
+    }
+    )"
+};
+
+/* Statement related to files items. Defines everything necessary to perform the synchronization loop */
+constexpr auto FIM_FILE_START_CONFIG_STATEMENT
+{
+    R"({"table":"file_entry"})"
+    //TO DO
+};
+
+/* Statement related to registries items. Defines everything necessary to perform the synchronization loop */
+constexpr auto FIM_REGISTRY_START_CONFIG_STATEMENT
+{
+    R"({"table":"registry_key"})"
+    //TO DO
+};
+
+/* Statement related to values items. Defines everything necessary to perform the synchronization loop */
+constexpr auto FIM_VALUE_START_CONFIG_STATEMENT
+{
+    R"({"table":"registry_data"})"
+    //TO DO
+};
+
 
 class FIMDB
 {
@@ -52,22 +241,24 @@ class FIMDB
          * @param max_rows_registry Maximun number of registry values entries in database (only for Windows)
          * @param callbackSync Pointer to the callback used to send sync messages
          * @param callbackLog Pointer to the callback used to send log messages
+         * @param dbsyncHandler Pointer to a dbsync handler.
+         * @param rsyncHandler Pointer to a rsync handler
          */
 #ifdef WIN32
         void init(unsigned int interval_synchronization,
                   unsigned int max_rows_file,
                   unsigned int max_rows_registry,
-                  send_data_callback_t callbackSync,
+                  fim_sync_callback_t callbackSync,
                   logging_callback_t callbackLog,
-                  std::unique_ptr<DBSync> dbsyncHandler,
-                  std::unique_ptr<RemoteSync> rsyncHandler);
+                  std::shared_ptr<DBSync> dbsyncHandler,
+                  std::shared_ptr<RemoteSync> rsyncHandler);
 #else
         void init(unsigned int interval_synchronization,
                   unsigned int max_rows_file,
-                  send_data_callback_t callbackSync,
+                  fim_sync_callback_t callbackSync,
                   logging_callback_t callbackLog,
-                  std::unique_ptr<DBSync> dbsyncHandler,
-                  std::unique_ptr<RemoteSync> rsyncHandler);
+                  std::shared_ptr<DBSync> dbsyncHandler,
+                  std::shared_ptr<RemoteSync> rsyncHandler);
 #endif
         /**
          * @brief Insert a given item into the database
@@ -127,8 +318,8 @@ class FIMDB
         unsigned int                                                            m_interval_synchronization;
         bool                                                                    m_stopping;
         std::condition_variable                                                 m_cv;
-        std::unique_ptr<DBSync>                                                 m_dbsyncHandler;
-        std::unique_ptr<RemoteSync>                                             m_rsyncHandler;
+        std::shared_ptr<DBSync>                                                 m_dbsyncHandler;
+        std::shared_ptr<RemoteSync>                                             m_rsyncHandler;
         std::function<void(const std::string&)>                                 m_syncMessageFunction;
         std::function<void(modules_log_level_t, const std::string&)>            m_loggingFunction;
