+++ src/syscheckd/db/src/fimDB.cpp
@@ -119,24 +119,24 @@ void FIMDB::loopRSync(std::unique_lock<std::mutex>& lock)
     {
         sync();
     }
-    m_rsyncHandler.reset(nullptr);
+    m_rsyncHandler = nullptr;
 }
 
 #ifdef WIN32
 void FIMDB::init(unsigned int interval_synchronization,
                  unsigned int max_rows_file,
                  unsigned int max_rows_registry,
-                 send_data_callback_t callbackSync,
+                 fim_sync_callback_t callbackSync,
                  logging_callback_t callbackLog,
-                 std::unique_ptr<DBSync> dbsyncHandler,
-                 std::unique_ptr<RemoteSync> rsyncHanlder)
+                 std::shared_ptr<DBSync> dbsyncHandler,
+                 std::shared_ptr<RemoteSync> rsyncHanlder)
 #else
 void FIMDB::init(unsigned int interval_synchronization,
                  unsigned int max_rows_file,
-                 send_data_callback_t callbackSync,
+                 fim_sync_callback_t callbackSync,
                  logging_callback_t callbackLog,
-                 std::unique_ptr<DBSync> dbsyncHandler,
-                 std::unique_ptr<RemoteSync> rsyncHanlder)
+                 std::shared_ptr<DBSync> dbsyncHandler,
+                 std::shared_ptr<RemoteSync> rsyncHanlder)
 #endif
 {
     // LCOV_EXCL_START
@@ -162,8 +162,8 @@ void FIMDB::init(unsigned int interval_synchronization,
 #ifdef WIN32
     m_max_rows_registry = max_rows_registry;
 #endif
-    m_dbsyncHandler = std::move(dbsyncHandler);
-    m_rsyncHandler = std::move(rsyncHanlder);
+    m_dbsyncHandler = dbsyncHandler;
+    m_rsyncHandler = rsyncHanlder;
     m_syncMessageFunction = callbackSyncWrapper;
     m_loggingFunction = callbackLogWrapper;
     m_stopping = false;
